# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LeyInvest API Documentation",
        "description": "Documentation complète de l'API LeyInvest. Cette API permet la gestion des utilisateurs, des actions suivies (follow/unfollow), ainsi que des opérations financières automatisées (take profit, stop loss).",
        "contact": {
            "name": "API Support",
            "email": "tiafranck@leycom.ci"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur principal de l'API LeyInvest"
        }
    ],
    "paths": {
...
        "/api/v1/user/actions": {
            "get": {
                "tags": [
                    "User Actions"
                ],
                "summary": "Lister les actions suivies par l'utilisateur authentifié",
                "description": "Récupère toutes les actions suivies par l'utilisateur authentifié",
                "operationId": "getFollowedActions",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserActionResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user/action/follow": {
            "post": {
                "tags": [
                    "User Actions"
                ],
                "summary": "Suivre une action",
                "description": "Permet à l'utilisateur de suivre une action avec des valeurs optionnelles de stop_loss et take_profit",
                "operationId": "followAction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FollowActionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Action suivie avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserActionResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "409": {
                        "description": "Action déjà suivie"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user/action/{actionId}/unfollow": {
            "delete": {
                "tags": [
                    "User Actions"
                ],
                "summary": "Ne plus suivre une action",
                "description": "Permet de ne plus suivre une action précédemment suivie",
                "operationId": "unfollowAction",
                "parameters": [
                    {
                        "name": "actionId",
                        "in": "path",
                        "description": "ID de l'action à unfollow",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Action retirée avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Action non trouvée ou non suivie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user/action/{actionId}": {
            "patch": {
                "tags": [
                    "User Actions"
                ],
                "summary": "Mettre à jour les paramètres de suivi",
                "description": "Permet à l'utilisateur de mettre à jour les valeurs stop_loss et take_profit d'une action déjà suivie",
                "operationId": "updateFollowParameters",
                "parameters": [
                    {
                        "name": "actionId",
                        "in": "path",
                        "description": "ID de l'action suivie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 8
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserActionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paramètres mis à jour avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Action non suivie"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user/dashboard": {
            "get": {
                "tags": [
                    "UserDashboard"
                ],
                "summary": "Récupérer le tableau de bord de l'utilisateur connecté",
                "description": "Retourne les informations du user connecté, 5 actions suivies aléatoires et sa liste complète d'actions suivies.",
                "operationId": "getUserDashboard",
                "responses": {
                    "200": {
                        "description": "Dashboard utilisateur récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserDashboardResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "FollowActionRequest": {
                "required": [
                    "action_id"
                ],
                "properties": {
                    "action_id": {
                        "description": "ID de l'action à suivre",
                        "type": "integer",
                        "example": 1
                    },
                    "stop_loss": {
                        "description": "Valeur optionnelle de stop loss",
                        "type": "number",
                        "format": "float",
                        "example": 950.5,
                        "nullable": true
                    },
                    "take_profit": {
                        "description": "Valeur optionnelle de take profit",
                        "type": "number",
                        "format": "float",
                        "example": 1200,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateUserActionRequest": {
                "properties": {
                    "stop_loss": {
                        "description": "Nouvelle valeur de stop loss",
                        "type": "number",
                        "format": "float",
                        "example": 980,
                        "nullable": true
                    },
                    "take_profit": {
                        "description": "Nouvelle valeur de take profit",
                        "type": "number",
                        "format": "float",
                        "example": 1250,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UserActionResource": {
                "title": "User Action Resource",
                "description": "Relation entre un utilisateur et une action suivie, incluant les paramètres stop_loss et take_profit",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 15
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "action_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "stop_loss": {
                        "type": "number",
                        "format": "float",
                        "example": 980,
                        "nullable": true
                    },
                    "take_profit": {
                        "type": "number",
                        "format": "float",
                        "example": 1250,
                        "nullable": true
                    },
                    "action": {
                        "description": "Informations sur l'action suivie",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "symbole": {
                                "type": "string",
                                "example": "ABJC"
                            },
                            "nom": {
                                "type": "string",
                                "example": "SERVAIR ABIDJAN COTE D'IVOIRE"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "user": {
                        "description": "Informations sur l'utilisateur qui suit cette action",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 3
                            },
                            "nom": {
                                "type": "string",
                                "example": "Kouassi"
                            },
                            "prenom": {
                                "type": "string",
                                "example": "Eren"
                            },
                            "email": {
                                "type": "string",
                                "example": "eren@example.com"
                            },
                            "numero": {
                                "type": "string",
                                "example": "+2250700000000"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-21T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-21T13:10:00Z"
                    }
                },
                "type": "object"
            },
            "UserDashboardResource": {
                "title": "Ressource du tableau de bord utilisateur",
                "description": "Structure de données renvoyée par l'API /api/v1/user/dashboard",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "numero": {
                        "type": "string",
                        "example": "+2250102030405"
                    },
                    "actions_suivis": {
                        "description": "5 actions suivies aléatoirement",
                        "type": "array",
                        "items": {
                            "properties": {
                                "key": {
                                    "type": "string",
                                    "example": "ABJC"
                                },
                                "symbole": {
                                    "type": "string",
                                    "example": "SERVAIR ABIDJAN CÔTE D’IVOIRE"
                                },
                                "nom": {
                                    "type": "string",
                                    "example": "SERVAIR ABIDJAN CÔTE D’IVOIRE"
                                },
                                "volume": {
                                    "type": "number",
                                    "example": 5759
                                },
                                "cours_veille": {
                                    "type": "number",
                                    "example": 980
                                },
                                "cours_ouverture": {
                                    "type": "number",
                                    "example": 985
                                },
                                "cours_cloture": {
                                    "type": "number",
                                    "example": 1005.5
                                },
                                "variation": {
                                    "type": "number",
                                    "example": 2.4
                                },
                                "couleur_variation": {
                                    "type": "string",
                                    "example": "green"
                                },
                                "stop_loss": {
                                    "type": "number",
                                    "example": 950,
                                    "nullable": true
                                },
                                "take_profit": {
                                    "type": "number",
                                    "example": 1250,
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    },
                    "ma_liste": {
                        "description": "Liste complète de toutes les actions suivies",
                        "type": "array",
                        "items": {
                            "properties": {
                                "key": {
                                    "type": "string",
                                    "example": "BOAC"
                                },
                                "symbole": {
                                    "type": "string",
                                    "example": "BANK OF AFRICA CÔTE D’IVOIRE"
                                },
                                "nom": {
                                    "type": "string",
                                    "example": "BANK OF AFRICA CÔTE D’IVOIRE"
                                },
                                "volume": {
                                    "type": "number",
                                    "example": 4562
                                },
                                "cours_veille": {
                                    "type": "number",
                                    "example": 900
                                },
                                "cours_ouverture": {
                                    "type": "number",
                                    "example": 920
                                },
                                "cours_cloture": {
                                    "type": "number",
                                    "example": 910
                                },
                                "variation": {
                                    "type": "number",
                                    "example": -1.1
                                },
                                "stop_loss": {
                                    "type": "number",
                                    "example": 850,
                                    "nullable": true
                                },
                                "take_profit": {
                                    "type": "number",
                                    "example": 1150,
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token Bearer obtenu via l'authentification Sanctum (ex: 'Bearer {token}')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}